script_telnet/requirements.txt
python-dotenv==1.2.1
telnetlib3==2.2.0
